Added functionality to create a credit transaction

+++ library/src/main/java/com/thoughtworks/step/bank/Transaction.java
@@ -5,10 +5,10 @@ import java.util.Objects;
 
 public abstract class Transaction {
     protected Date date;
-    protected final float amount;
+    protected final double amount;
     protected final String to;
 
-    public Transaction(Date date, float amount, String to) {
+    public Transaction(Date date, double amount, String to) {
         this.date = date;
         this.amount = amount;
         this.to = to;
@@ -23,7 +23,7 @@ public abstract class Transaction {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
         Transaction that = (Transaction) o;
-        return Float.compare(that.amount, amount) == 0 &&
+        return Double.compare(that.amount, amount) == 0 &&
                 Objects.equals(to, that.to);
     }
